Move authModel to AuthController
