+++ src/api/models/index.js
@@ -1,2 +1,1 @@
-export { default as authModel } from './authModel';
 export { default as userModel } from './userModel';
